+++ src/view/announcer/announcer.js
@@ -3,7 +3,7 @@ import invariant from 'tiny-invariant';
 import type { Announce } from '../../types';
 import type { Announcer } from './announcer-types';
 import { warning } from '../../dev-warning';
-import getBody from '../get-body';
+import getBodyElement from '../get-body-element';
 
 let count: number = 0;
 
@@ -63,14 +63,14 @@ export default (): Announcer => {
     Object.assign(el.style, visuallyHidden);
 
     // Add to body
-    getBody().appendChild(el);
+    getBodyElement().appendChild(el);
   };
 
   const unmount = () => {
     invariant(el, 'Will not unmount announcer as it is already unmounted');
 
     // Remove from body
-    getBody().removeChild(el);
+    getBodyElement().removeChild(el);
     // Unset
     el = null;
   };
