+++ src/view/get-body-element.js
