+++ src/view/get-document-element.js
