+++ src/view/window/get-max-window-scroll.js
@@ -1,10 +1,10 @@
 // @flow
 import type { Position } from 'css-box-model';
 import getMaxScroll from '../../state/get-max-scroll';
-import getDocument from '../get-document';
+import getDocumentElement from '../get-document-element';
 
 export default (): Position => {
-  const doc: HTMLElement = getDocument();
+  const doc: HTMLElement = getDocumentElement();
 
   const maxScroll: Position = getMaxScroll({
     // unclipped padding box, with scrollbar
