+++ src/view/window/get-viewport.js
@@ -4,7 +4,7 @@ import type { Viewport } from '../../types';
 import { origin } from '../../state/position';
 import getWindowScroll from './get-window-scroll';
 import getMaxWindowScroll from './get-max-window-scroll';
-import getDocument from '../get-document';
+import getDocumentElement from '../get-document-element';
 
 export default (): Viewport => {
   const scroll: Position = getWindowScroll();
@@ -15,7 +15,7 @@ export default (): Viewport => {
 
   // window.innerHeight: includes scrollbars (not what we want)
   // document.clientHeight gives us the correct value when using the html5 doctype
-  const doc: HTMLElement = getDocument();
+  const doc: HTMLElement = getDocumentElement();
   // Using these values as they do not consider scrollbars
   // padding box, without scrollbar
   const width: number = doc.clientWidth;
